Fix unresolved references errors in GoLand (#52)
This commit fixes the following issues listed in #52:

  * Unresolved references due to the variables `folder` and
    `environmentsFile` being defined in file
    `multi_tenant_integration_test.go`. They can be seen when opening
    the project in GoLand.
  * We create configs for different environments, as we try to clean
    after running the integration tests.
  * We try to clean up different configs, as we apply in the
    integration test run.

Furthermore, the method `RunIntegrationWithCleanup` was simplified to
take a `func(integrationTestReader util.FileReader) {}`. This
simplifies the integration tests, because the `InMemoryFileReader` is
created in the utility function and passed to the test itself. So, the
individual test doesn't need to create a `FileReader` anymore.

Additionally, some documentation was added to the function
`RunIntegrationWithCleanup`.